backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pyyaml>=6.0
pillow>=10.0.0
loguru>=0.7.0
rich>=13.0.0
aiofiles>=23.0.0
python-multipart>=0.0.6
